requirements.txt
# FinBundle Dependencies
# Vector DB & Embeddings
qdrant-client>=1.13.0  # ACORN filtered search + text_any support
sentence-transformers>=2.2.0

# LLM
groq>=0.4.0

# Optimization
ortools>=9.7.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=8.3.4

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Web Framework - API
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0

# Legacy Streamlit (optional)
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
tqdm>=4.65.0

# ML/Deep Learning
torch>=2.0.0
transformers>=4.35.0

